refactor: proxy 도메인 계층 분리
- handler에서 프록시 도메인 로직을 service로 이동
- 프록시 URL 해석 로직을 service로 분리
- handler는 요청/응답 처리만 담당하도록 변경
